Finally, at long last, the problem with the applet dying when the page refreshed has been solved. This endeavor required a deep descent into the dark depths of the Java Monkey Engine. The solution to this problem partially involves a drastic change in which the applet is loaded.
Instead of using the default jme distribution jars, we instead build our own and sign it ourselves. This allows us to make necessary changes in the jME code, as well as reducing by one the number of confirmation popups.

The .keystore file has also been updated, so signed jars will now no longer give errors that the signature is invalid.